Backend/requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database & Vector Store
pinecone-client==2.2.4
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==5.0.1

# LLM & AI
openai==1.3.0
anthropic==0.7.7
sentence-transformers==2.2.2
langchain==0.0.340
langchain-openai==0.0.1

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
aiohttp==3.9.1
beautifulsoup4==4.12.2

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk==1.38.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
celery==5.3.4
websockets==12.0
python-socketio==5.10.0